requirements.txt
BeautifulSoup==3.2.1
Django==1.4.5
Fabric==1.5.3
GitPython==0.3.2.RC1
South==0.7.6
Werkzeug==0.8.3
async==0.6.1
boto==2.8.0
dj-database-url==0.2.1
django-appconf==0.6
django-compressor==1.2
django-extensions==1.0.3
django-nose==1.1
django-storages==1.1.6
django-waffle==0.9.0
djangorestframework==2.1.17
gitdb==0.5.4
lxml==3.1.2
nose==1.3.0
paramiko==1.10.1
psycopg2==2.4.5
pycrypto==2.6
requests==1.2.0
selenium==2.32.0
six==1.3.0
smmap==0.8.2
splinter==0.5.0
django-celery-with-redis==3.0
wsgiref==0.1.2
